+++ libs/platform/core/src/services/page-meta/page-meta-resolver/index.ts
@@ -1,3 +1,4 @@
 export * from './default-page-meta-resolver.service';
 export * from './global-page-meta-resolver';
+export * from './page-meta-resolver.model';
 export * from './page-meta-resolver.service';
