+++ libs/platform/core/src/services/page-meta/page-meta-resolver/page-meta-resolver.service.ts
@@ -1,25 +1,17 @@
 import { OnDestroy } from '@spryker-oryx/di';
 import { Observable } from 'rxjs';
 import { AppInitializer } from '../../app-initializer';
-import { ElementDefinition } from '../page-meta.model';
+import { ElementResolver, ResolverScore } from './page-meta-resolver.model';
 
 export const PageMetaResolverService = 'oryx.PageMetaResolverService';
 export const PageMetaResolver = 'oryx.PageMetaResolver*';
-
-export interface ElementResolver extends Partial<ElementDefinition> {
-  link?: string;
-  style?: string;
-  script?: string;
-  title?: string;
-  id?: string;
-}
 
 export interface PageMetaResolverService extends AppInitializer, OnDestroy {
   getTitle(): Observable<string | undefined>;
 }
 
 export interface PageMetaResolver {
-  getScore(): Observable<number>;
+  getScore(): Observable<ResolverScore | number>;
   resolve(): Observable<ElementResolver | ElementResolver[]>;
 }
